+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - Album preview in invite page
 - Upgrade to Next.js 14.1
+- Show modal for invite preview
 
 ## 2024-02-19
 - Move wedding link to header